+++ src/Kiota.Builder/CodeDOM/CodeFunction.cs
@@ -23,10 +23,19 @@ public class CodeFunction : CodeBlock<BlockDeclaration, BlockEnd>
     {
         get; private set;
     }
+    public CodeClass OriginalMethodParentClass
+    {
+        get;
+        private set;
+    }
     public CodeFunction(CodeMethod method)
     {
-        if (method == null) throw new ArgumentNullException(nameof(method));
+        ArgumentNullException.ThrowIfNull(method);
         if (!method.IsStatic) throw new InvalidOperationException("The original method must be static");
+        if (method.Parent is CodeClass parentClass)
+            OriginalMethodParentClass = parentClass;
+        else
+            throw new InvalidOperationException("The original method must be a member of a class");
         EnsureElementsAreChildren(method);
         OriginalLocalMethod = method;
     }