+++ src/Kiota.Builder/CodeDOM/CodeIntersectionType.cs
@@ -5,10 +5,10 @@ namespace Kiota.Builder;
 /// <summary>
 /// The base class for exclusion types. (one of the properties at a time)
 /// </summary>
-public class CodeExclusionType : CodeComposedTypeBase, ICloneable {
+public class CodeIntersectionType : CodeComposedTypeBase, ICloneable {
     public override object Clone() {
-        var value = new CodeExclusionType{
-        }.BaseClone<CodeExclusionType>(this);
+        var value = new CodeIntersectionType{
+        }.BaseClone<CodeIntersectionType>(this);
         return value;
     }
 }